OCS/requirements.txt
# Core
Django==5.1.1

# Data Processing
pandas==2.2.3
numpy==2.1.1
openpyxl==3.1.5

# Configuration
python-decouple==3.8

# Production Server
gunicorn==22.0.0
whitenoise==6.7.0

# Django Dependencies
asgiref==3.8.1
sqlparse==0.5.1
tzdata==2024.2

# Pandas Dependencies
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
et-xmlfile==1.1.0

# Database Connectors (optional — only needed if using MySQL)
# mysqlclient==2.2.4
# mysql-connector-python==8.4.0

# Testing
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0

# Code Quality
flake8==7.1.1
black==24.8.0
isort==5.13.2
